Add refactor for ContributionMetaBar
