+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {Route} from 'react-router-dom';
 import {connect} from 'react-redux';
 import {withRouter} from 'react-router-dom';
-import {Layout} from 'antd';
+import {Layout,Col} from 'antd';
 import KnackHeader from './components/Header';
 import Home from './containers/Home';
 import Guidelines from './containers/Guidelines';
@@ -25,12 +25,14 @@ import './assets/styles/ant-overrides.css';
 import Sponsors from './containers/Sponsors';
 import Moderators from './containers/Moderators';
 import Search from './containers/Search';
+import ContributionMetaBar from './containers/ContributionMetaBar'
 const App = () => {
   return (
     <Layout id="page-layout">
       <KnackHeader/>
       <Route exact path="/@:username" component={Profile} />
       <Layout id="content-layout">
+        <ContributionMetaBar/>
         <Route exact path="/" component={Home} />
         <Route exact path="/guidelines" component={Guidelines} />
         <Route exact path="/faq" component={FAQs} />
