+++ src/containers/Home/index.css
@@ -1,7 +1,6 @@
 .home-container {
     justify-content: center;
     width: 100%;
-    margin-top: 100px
 }
 .contribution-container,
 .announcement-container{
