+++ src/containers/Home/index.js
@@ -6,7 +6,7 @@ import './index.css';
 import {Layout, Spin, Row, Col, Select } from 'antd';
 import ArticleListItem from '../../components/ArticleListItem';
 import AnnouncementMetaBar from '../Home/AnnouncementMetaBar';
-import ContributionMetaBar from '../Home/ContributionMetaBar';
+import ContributionMetaBar from '../ContributionMetaBar';
 import { getArticlesModeration} from '../../actions/articles';
 import {getRemoteUserData} from '../../actions/user';
 import Cookies from 'js-cookie';
@@ -33,7 +33,6 @@ class Home extends Component {
     this.state = {
       searchString: ''
     };
-    this.getOathURL = this.getOathURL.bind(this);
   }
 
   // //scroll handler for lazy loading
@@ -51,17 +50,8 @@ class Home extends Component {
      }
    };
   //get User data
-   loadRemoteUserData() {
-     const {dispatch} = this.props;
-     dispatch(getRemoteUserData(Cookies.get('username')));
-   }
+   
 
-   /**
-    * @method st
-    */
-   getOathURL () {
-     return SteemConnect.getLoginURL();
-   }
 
    componentDidMount() {
      const username = Cookies.get('username');
@@ -69,11 +59,11 @@ class Home extends Component {
        this.loadArticles();
      } else {
        this.loadArticles();
-       this.loadRemoteUserData();
      }
      // //on scroll, load the next batch of articles
      window.addEventListener('scroll', this.onScroll);
    }
+
    componentWillUnmount() {
      //remove scroll event again when hitting another route
      window.removeEventListener('scroll', this.onScroll);
@@ -95,24 +85,9 @@ class Home extends Component {
   };
 
   render() {
-    let 
-      coverImage,
-      name,
-      reputation,
-      remoteUserObjectMeta,
-      username;
 
-    const {articles, user} = this.props;
-    const { remoteUserObject} = user;
-    const hasLoadedRemoteUserObject = Object.keys(remoteUserObject).length > 0;
-
-    if (hasLoadedRemoteUserObject) {
-      remoteUserObjectMeta = JSON.parse(remoteUserObject.json_metadata).profile;
-      name = remoteUserObjectMeta.name;
-      coverImage = remoteUserObjectMeta.cover_image;
-      reputation = repLog10(parseFloat(remoteUserObject.reputation)); 
-      username = Cookies.get('username');
-    }  
+    const {articles} = this.props;
+ 
 
     return (
       <Layout className="home-container" justify="center">
@@ -142,11 +117,6 @@ class Home extends Component {
             
         </Row>
         <Row type="flex" className="home-inner-container" justify="center">
-          <Row type="flex" className="contribution-container">
-            <Col>
-              <ContributionMetaBar metaImage={coverImage} reputation={reputation} name={name}  handleLogin={this.getOathURL()} username={username}/>
-            </Col>
-          </Row>
           <Row className="item-feed ant-list ant-list-vertical ant-list-lg ant-list-split ant-list-something-after-last-item" style={styles.articlesList}>
             {articles.data.map((data) => {
               return (
